refactor(campaigns): extract shared helpers to _campaigns_base.py (#602, step 1)
First step of an incremental split of the 7 209-line campaigns.py monolith
(issue #602). Moves all campaign-type-agnostic code into a new sibling module
direct_cli/commands/_campaigns_base.py:

- constants (CAMPAIGNS_GET_CRITERIA_LIMITS, YES_NO, ATTRIBUTION_MODELS, ...)
- typed-flag validators (_validate_max_length, _validate_sms_time)
- payload builders (_build_notification, _build_time_targeting,
  _build_relevant_keywords, _build_dynamic_placement_types,
  _build_frequency_cap, _build_package_bidding_strategy,
  _build_smart_package_bidding_strategy, _priority_goals_update_items,
  _route_cpa_flag)
- reusable TextCampaign strategy click.option groups
  (_TEXT_*_STRATEGY_OPTIONS and their *_UPDATE variants) + _apply_options

campaigns.py re-imports every name, so the CLI surface, every flag and every
--dry-run payload is byte-for-byte identical (offline tier 2507 passed,
test_wsdl_parity_gate + test_api_coverage green). Per-campaign-type logic
(text/unified/dynamic/smart/mobile_app/cpm_banner) follows in subsequent PRs.

Closes part of #602.
